implement env-clone command
This commit is implemented to support the release procedure, described at
http://support.cloudfoundry.com/entries/20022057-how-to-update-your-application-without-dropping-user-traffic

In the above procedure, they does not mention about the service binding and environment variables of the old application, but the basic strategy would be 'to copy settings from the old to the new'.

Assuming this basic strategy, service bindings can be copied by 'clone-services' whereas environment variables cannot. So this patch would add the opeartion 'to copy environment variables from the old app to the new'.

With this commit, the release procedure should be described as follows:

    $ vmc push new-app --no-start -n
    $ vmc clone-sevices old-app new-app
    $ vmc env-clone old-app new-app

    # configure services, environment variablse if needed.

    $ vmc start new-app

    # test application

    $ vmc map new-app appurl
    $ vmc unmap old-app appurl